Update to SDK 2.0 TimeZone now stored persistently removed obsolete unix epoch function moved sha stuff into separate files, merged editTzone.c into authenicator.c
