--- resources/src/resource_map.json
@@ -1,10 +1,0 @@
-
-{"friendlyVersion": "VERSION",
- "versionDefName": "VERSION",
- "media": [
-	{  "type": "png",
-	   "defName": "IMAGE_MENU_ICON",
-	   "file": "images/twostep_icon.png"
-	}     	   
-	 ]
-}
